[PR196 3/3] Adding support for partial workspace cleaning

+++ catkin_tools/jobs/catkin.py
@@ -14,7 +14,6 @@
 
 import csv
 import os
-import threading
 
 try:
     from md5 import md5
@@ -32,14 +31,16 @@ from catkin_tools.execution.stages import FunctionStage
 from .commands.cmake import CMAKE_EXEC
 from .commands.cmake import CMakeIOBufferProtocol
 from .commands.cmake import CMakeMakeIOBufferProtocol
+from .commands.cmake import get_installed_files
 from .commands.make import MAKE_EXEC
 
+from .utils import copyfiles
 from .utils import get_env_loader
 from .utils import makedirs
-from .utils import rmdirs
+from .utils import rmfiles
 
 
-def generate_prebuild_package(build_space_abs, devel_space_abs, force):
+def get_prebuild_package(build_space_abs, devel_space_abs, force):
     """This generates a minimal Catkin package used to generate Catkin
     environment setup files in a merged devel space.
 
@@ -50,7 +51,7 @@ def generate_prebuild_package(build_space_abs, devel_space_abs, force):
     """
 
     # Get the path to the prebuild package
-    prebuild_path = os.path.join(devel_space_abs, '.catkin_tools', 'catkin_tools_prebuild')
+    prebuild_path = os.path.join(devel_space_abs, '.private', 'catkin_tools_prebuild')
     if not os.path.exists(prebuild_path):
         mkdir_p(prebuild_path)
 
@@ -75,9 +76,10 @@ def generate_prebuild_package(build_space_abs, devel_space_abs, force):
 def clean_linked_files(
         logger,
         event_queue,
-        devel_space_abs,
+        metadata_path,
         files_that_collide,
-        files_to_clean):
+        files_to_clean,
+        dry_run):
     """Removes a list of files and adjusts collison counts for colliding files.
 
     This function synchronizes access to the devel collisions file.
@@ -88,7 +90,7 @@ def clean_linked_files(
     """
 
     # Get paths
-    devel_collisions_file_path = os.path.join(devel_space_abs, 'devel_collisions.txt')
+    devel_collisions_file_path = os.path.join(metadata_path, 'devel_collisions.txt')
 
     # Map from dest files to number of collisions
     dest_collisions = dict()
@@ -113,14 +115,22 @@ def clean_linked_files(
 
         # Check collisions
         if n_collisions == 0:
-            logger.out('Unlinking %s' % (dest_file))
+            logger.out('Unlinking: {}'.format(dest_file))
             # Remove this link
-            os.unlink(dest_file)
-            # Remove any non-empty directories containing this file
-            try:
-                os.removedirs(os.path.split(dest_file)[0])
-            except OSError:
-                pass
+            if not dry_run:
+                if os.path.exists(dest_file):
+                    try:
+                        os.unlink(dest_file)
+                    except OSError:
+                        logger.err('Could not unlink: {}'.format(dest_file))
+                        raise
+                    # Remove any non-empty directories containing this file
+                    try:
+                        os.removedirs(os.path.split(dest_file)[0])
+                    except OSError:
+                        pass
+                else:
+                    logger.out('Already unlinked: {}')
 
         # Update collisions
         if n_collisions > 1:
@@ -131,31 +141,40 @@ def clean_linked_files(
             del dest_collisions[dest_file]
 
     # Load destination collisions file
-    with open(devel_collisions_file_path, 'w') as collisions_file:
-        collisions_writer = csv.writer(collisions_file, delimiter=' ', quotechar='"')
-        for dest_file, count in dest_collisions.items():
-            collisions_writer.writerow([dest_file, count])
+    if not dry_run:
+        with open(devel_collisions_file_path, 'w') as collisions_file:
+            collisions_writer = csv.writer(collisions_file, delimiter=' ', quotechar='"')
+            for dest_file, count in dest_collisions.items():
+                collisions_writer.writerow([dest_file, count])
 
 
 def unlink_devel_products(
         logger,
         event_queue,
         devel_space_abs,
-        package_name):
+        private_devel_path,
+        metadata_path,
+        package_metadata_path,
+        dry_run):
     """
     Remove all files listed in the devel manifest for the given package, as
     well as any empty directories containing those files.
 
     :param devel_space_abs: Path to a merged devel space.
-    :param package_name: Name of the package whose files should be unlinked.
+    :param private_devel_path: Path to the private devel space
+    :param devel_manifest_path: Path to the directory containing the package's
+    catkin_tools metadata
     """
 
-    # Get paths
-    linked_devel_path = get_linked_devel_package_path(devel_space_abs, package_name)
-    devel_manifest_path = get_linked_devel_manifest_path(devel_space_abs, package_name)
-
-    if not os.path.exists(linked_devel_path) or not os.path.exists(devel_manifest_path):
+    # Check paths
+    if not os.path.exists(private_devel_path):
+        logger.err('Warning: No private devel path found at `{}`'.format(private_devel_path))
         return 0
+
+    devel_manifest_file_path = os.path.join(package_metadata_path, DEVEL_MANIFEST_FILENAME)
+    if not os.path.exists(devel_manifest_file_path):
+        logger.err('Error: No devel manifest found at `{}`'.format(devel_manifest_file_path))
+        return 1
 
     # List of files to clean
     files_to_clean = []
@@ -181,7 +200,7 @@ def unlink_devel_products(
 
     # Remove all listed symli and empty directories which have been removed
     # after this build, and update the collision file
-    clean_linked_files(logger, event_queue, devel_space_abs, [], files_to_clean)
+    clean_linked_files(logger, event_queue, metadata_path, [], files_to_clean, dry_run)
 
     return 0
 
@@ -192,12 +211,17 @@ def link_devel_products(
         package_path,
         devel_manifest_path,
         source_devel_path,
-        dest_devel_path):
+        dest_devel_path,
+        metadata_path,
+        prebuild):
     """Link files from an isolated devel space into a merged one.
 
     This creates directories and symlinks in a merged devel space to a
     package's linked devel space.
     """
+
+    # Create the devel manifest path if necessary
+    mkdir_p(devel_manifest_path)
 
     # Construct manifest file path
     devel_manifest_file_path = os.path.join(devel_manifest_path, DEVEL_MANIFEST_FILENAME)
@@ -208,6 +232,9 @@ def link_devel_products(
     files_to_clean = []
     # List of files that collide
     files_that_collide = []
+
+    # Select the blacklist
+    blacklist = DEVEL_LINK_PREBUILD_BLACKLIST if prebuild else DEVEL_LINK_BLACKLIST
 
     # Gather all of the files in the devel space
     for source_path, dirs, files in os.walk(source_devel_path):
@@ -229,8 +256,8 @@ def link_devel_products(
         # create symbolic links from the source to the dest
         for filename in files:
 
-            # Don't link files on the blacklist
-            if os.path.relpath(os.path.join(source_path, filename), source_devel_path) in devel_link_blacklist:
+            # Don't link files on the blacklist unless this is a prebuild package
+            if os.path.relpath(os.path.join(source_path, filename), source_devel_path) in blacklist:
                 continue
 
             source_file = os.path.join(source_path, filename)
@@ -253,10 +280,16 @@ def link_devel_products(
                         logger.err('Warning: Dest hash: {}'.format(dest_hash))
                     # Increment link collision counter
                     files_that_collide.append(dest_file)
+                else:
+                    logger.out('Linked: ({}, {})'.format(source_file, dest_file))
             else:
                 # Create the symlink
                 logger.out('Symlinking %s' % (dest_file))
-                os.symlink(source_file, dest_file)
+                try:
+                    os.symlink(source_file, dest_file)
+                except OSError:
+                    logger.err('Could not create symlink `{}` referencing `{}`'.format(dest_file, source_file))
+                    raise
 
     # Load the old list of symlinked files for this package
     if os.path.exists(devel_manifest_file_path):
@@ -269,12 +302,16 @@ def link_devel_products(
                 # print('Checking (%s, %s)' % (source_file, dest_file))
                 if (source_file, dest_file) not in products:
                     # Clean the file or decrement the collision count
-                    logger.out('Cleaning (%s, %s)' % (source_file, dest_file))
+                    logger.out('Cleaning: (%s, %s)' % (source_file, dest_file))
                     files_to_clean.append(dest_file)
 
     # Remove all listed symlinks and empty directories which have been removed
     # after this build, and update the collision file
-    clean_linked_files(logger, event_queue, dest_devel_path, files_that_collide, files_to_clean)
+    try:
+        clean_linked_files(logger, event_queue, metadata_path, files_that_collide, files_to_clean, dry_run=False)
+    except:
+        logger.err('Could not clean linked files.')
+        raise
 
     # Save the list of symlinked files
     with open(devel_manifest_file_path, 'w') as devel_manifest:
@@ -309,12 +346,11 @@ def create_catkin_build_job(context, package, package_path, dependencies, force_
     # Package build space path
     build_space = context.package_build_space(package)
     # Package devel space path
-    if prebuild:
-        devel_space = context.devel_space_abs
-    else:
-        devel_space = context.package_devel_space(package)
+    devel_space = context.package_devel_space(package)
     # Package install space path
     install_space = context.package_install_space(package)
+    # Package metadata path
+    metadata_path = context.package_metadata_path(package)
 
     # Create job stages
     stages = []
@@ -324,6 +360,21 @@ def create_catkin_build_job(context, package, package_path, dependencies, force_
         'mkdir',
         makedirs,
         path=build_space
+    ))
+
+    # Create package metadata dir
+    stages.append(FunctionStage(
+        'mkdir',
+        makedirs,
+        path=metadata_path
+    ))
+
+    # Copy source manifest
+    stages.append(FunctionStage(
+        'cache-manifest',
+        copyfiles,
+        source_paths=[os.path.join(context.source_space_abs, package_path, 'package.xml')],
+        dest_path=os.path.join(metadata_path, 'package.xml')
     ))
 
     # Define test results directory
@@ -401,16 +452,18 @@ def create_catkin_build_job(context, package, package_path, dependencies, force_
     ))
 
     # Symlink command if using a linked develspace
-    if context.link_devel and not prebuild:
+    if context.link_devel:
         stages.append(FunctionStage(
             'symlink',
             link_devel_products,
             locked_resource='symlink-collisions-file',
             package=package,
             package_path=package_path,
-            devel_manifest_path=context.package_linked_devel_path(package),
+            devel_manifest_path=context.package_metadata_path(package),
             source_devel_path=context.package_devel_space(package),
-            dest_devel_path=context.devel_space_abs
+            dest_devel_path=context.devel_space_abs,
+            metadata_path=context.metadata_path(),
+            prebuild=prebuild
         ))
 
     # Make install command, if installing
@@ -430,73 +483,99 @@ def create_catkin_build_job(context, package, package_path, dependencies, force_
         stages=stages)
 
 
-def create_catkin_clean_job(context, package, package_path, dependencies):
+def create_catkin_clean_job(
+        context,
+        package,
+        package_path,
+        dependencies,
+        dry_run,
+        clean_build,
+        clean_devel,
+        clean_install):
     """Generate a Job that cleans a catkin package"""
 
     stages = []
 
-    # If the build space doesn't exist, do nothing
+    # Package build space path
     build_space = context.package_build_space(package)
-    if not os.path.exists(build_space):
-        return Job(jid=package_name, deps=dependencies, stages=[])
+    # Package metadata path
+    metadata_path = context.package_metadata_path(package)
 
-    # For isolated devel space, remove it entirely
-    if context.isolate_devel:
-        devel_space = os.path.join(context.devel_space_abs, package_name)
-
-        stages.append(FunctionStage('clean', rmdirs, path=devel_space))
-
-        return Job(
-            jid=package_name,
-            deps=dependencies,
-            stages=stages)
-
-    elif context.link_devel:
-        devel_space = os.path.join(build_space, 'devel')
-    else:
-        devel_space = context.devel_space_abs
-
-    # For isolated install space, remove it entirely
-    if context.isolate_install:
-        install_space = os.path.join(context.install_space_abs, package_name)
-
-        stages.append(FunctionStage('clean', rmdirs, path=install_space))
-
-        return Job(
-            jid=package_name,
-            deps=dependencies,
-            stages=stages)
-    else:
-        install_space = context.install_space_abs
-
-    if context.link_devel:
-        # Remove symlinked products
+    # Remove installed files
+    if clean_install:
+        installed_files = get_installed_files(context.package_metadata_path(package))
         stages.append(FunctionStage(
-            'unlink',
-            unlink_devel_products,
-            devel_space_abs=context.devel_space_abs,
-            package_name=package_name,
-            locked_resource='symlink-collisions-file'
-        ))
+            'cleaninstall',
+            rmfiles,
+            paths=sorted(installed_files),
+            remove_empty=True,
+            empty_root=context.install_space_abs,
+            dry_run=dry_run))
 
-        # Remove devel space
-        stages.append(FunctionStage(
-            'rmdevel', rmdirs,
-            path=context.package_devel_space(package)))
+    # Remove products in develspace
+    if clean_devel:
+        if context.merge_devel:
+            # Remove build targets from devel space
+            stages.append(CommandStage(
+                'clean',
+                [MAKE_EXEC, 'clean'],
+                cwd=build_space,
+            ))
+        elif context.link_devel:
+            # Remove symlinked products
+            stages.append(FunctionStage(
+                'unlink',
+                unlink_devel_products,
+                locked_resource='symlink-collisions-file',
+                devel_space_abs=context.devel_space_abs,
+                private_devel_path=context.package_private_devel_path(package),
+                metadata_path=context.metadata_path(),
+                package_metadata_path=context.package_metadata_path(package),
+                dry_run=dry_run
+            ))
+
+            # Remove devel space
+            stages.append(FunctionStage(
+                'rmdevel',
+                rmfiles,
+                paths=[context.package_private_devel_path(package)],
+                dry_run=dry_run))
+        elif context.isolate_devel:
+            # Remove devel space
+            stages.append(FunctionStage(
+                'rmdevel',
+                rmfiles,
+                paths=[context.package_devel_space(package)],
+                dry_run=dry_run))
 
     # Remove build space
-    stages.append(FunctionStage('rmbuild', rmdirs, path=build_space))
+    if clean_build:
+        stages.append(FunctionStage(
+            'rmbuild',
+            rmfiles,
+            paths=[build_space],
+            dry_run=dry_run))
+
+    # Remove cached metadata
+    if clean_build and clean_devel and clean_install:
+        stages.append(FunctionStage(
+            'rmmetadata',
+            rmfiles,
+            paths=[metadata_path],
+            dry_run=dry_run))
 
     return Job(
-        jid=package_name,
+        jid=package.name,
         deps=dependencies,
+        env_loader=get_env_loader(package, context),
         stages=stages)
 
 
 description = dict(
     build_type='catkin',
     description="Builds a catkin package.",
-    create_build_job=create_catkin_build_job
+    create_build_job=create_catkin_build_job,
+    create_clean_job=create_catkin_clean_job
 )
 
 
@@ -509,18 +588,19 @@ unset ROS_TEST_RESULTS_DIR
 DEVEL_MANIFEST_FILENAME = 'devel_manifest.txt'
 
 # List of files which shouldn't be copied
-devel_link_blacklist = [
+DEVEL_LINK_PREBUILD_BLACKLIST = [
+    '.catkin',
+    '.rosinstall',
+]
+DEVEL_LINK_BLACKLIST = DEVEL_LINK_PREBUILD_BLACKLIST + [
     os.path.join('etc', 'catkin', 'profile.d', '05.catkin_make.bash'),
     os.path.join('etc', 'catkin', 'profile.d', '05.catkin_make_isolated.bash'),
     os.path.join('etc', 'catkin', 'profile.d', '05.catkin-test-results.sh'),
-    '.catkin',
-    '.rosinstall',
     'env.sh',
     'setup.bash',
     'setup.zsh',
     'setup.sh',
     '_setup_util.py',
-    DEVEL_MANIFEST_FILENAME
 ]
 
 # CMakeLists.txt for prebuild package
